Add Quix Drive, drive mode select, drive encoders
Quix Drive is a fancy drive implementation for a wheel and throttle.
Other drive modes are plain Arcade Drive with the wheel/throttle and
Tank Drive with an Xbox controller.

Drive modes can be selected from a chooser on the SmartDashboard. The
default is Quix Drive.

Drive encoders are also being read from and exposed on the Drive module.